Location Gathering Fix and Screen Brightness
Fixed maps api controlling the poll rate and accuracy. Also, now that
this has been fixed, the location settings work correctly. Results are
now logged to "FixedLocation" table. Additionally, the screen brightness
is automatically set to "manual" mode and then forced to a value of 127
(from 0 to 255, so almost exactly 50%)